# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Medicare Advantage Plan Tracker API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /counties:
    get:
      operationId: listCounties
      tags: [counties]
      summary: List counties, optionally filtered by state
      parameters:
        - name: state_code
          in: query
          schema:
            type: string
          description: Two-letter state abbreviation
      responses:
        "200":
          description: List of counties
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/County"
  /counties/{fips}/plans:
    get:
      operationId: getCountyPlans
      tags: [counties]
      summary: Get all MA plans available in a specific county
      parameters:
        - name: fips
          in: path
          required: true
          schema:
            type: string
          description: 5-digit county FIPS code
        - name: plan_type
          in: query
          schema:
            type: string
            enum: [d_snp, i_snp, c_snp, regular]
          description: Filter by plan type
        - name: year
          in: query
          schema:
            type: number
      responses:
        "200":
          description: Plans in this county with enrollment data
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CountyPlan"
        "404":
          description: County not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /plans/{id}/benefits:
    get:
      operationId: getPlanBenefits
      tags: [benefits]
      summary: List attributed benefits for a plan
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      responses:
        "200":
          description: Benefits list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Benefit"
        "404":
          description: Plan not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    post:
      operationId: addPlanBenefit
      tags: [benefits]
      summary: Add a benefit to a plan
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BenefitInput"
      responses:
        "201":
          description: Benefit added
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Benefit"
        "404":
          description: Plan not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /plans/{id}/enrollments:
    get:
      operationId: getPlanEnrollments
      tags: [enrollments]
      summary: Get county-level enrollment data for a plan
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
        - name: year
          in: query
          schema:
            type: number
      responses:
        "200":
          description: Enrollment records by county
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Enrollment"
        "404":
          description: Plan not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /enrollments:
    get:
      operationId: listEnrollments
      tags: [enrollments]
      summary: List enrollment records across plans and counties
      parameters:
        - name: state_code
          in: query
          schema:
            type: string
        - name: county_fips
          in: query
          schema:
            type: string
        - name: plan_type
          in: query
          schema:
            type: string
            enum: [d_snp, i_snp, c_snp, regular]
        - name: year
          in: query
          schema:
            type: number
        - name: min_beneficiaries
          in: query
          schema:
            type: number
        - name: limit
          in: query
          schema:
            type: number
            default: 100
        - name: offset
          in: query
          schema:
            type: number
            default: 0
      responses:
        "200":
          description: Enrollment records
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Enrollment"
    post:
      operationId: createEnrollment
      tags: [enrollments]
      summary: Create an enrollment record
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EnrollmentInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Enrollment"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
    County:
      type: object
      properties:
        fips:
          type: string
        county_name:
          type: string
        state_code:
          type: string
        state_name:
          type: string
      required:
        - fips
        - county_name
        - state_code
        - state_name
...
    Benefit:
      type: object
      properties:
        id:
          type: number
        plan_id:
          type: number
        benefit_name:
          type: string
        benefit_category:
          type: string
          description: Category (dental, vision, hearing, otc, transportation, fitness, meals, etc.)
        benefit_value:
          type: ["string", "null"]
          description: Monetary value or description (e.g. "$1,500/year", "Included", "Not covered")
        is_attributed:
          type: boolean
          description: Whether this is an attributed supplemental benefit
      required:
        - id
        - plan_id
        - benefit_name
        - benefit_category
        - is_attributed
    BenefitInput:
      type: object
      properties:
        benefit_name:
          type: string
        benefit_category:
          type: string
        benefit_value:
          type: ["string", "null"]
        is_attributed:
          type: boolean
      required:
        - benefit_name
        - benefit_category
        - is_attributed
    Enrollment:
      type: object
      properties:
        id:
          type: number
        plan_id:
          type: number
        plan_name:
          type: string
        contract_id:
          type: string
        plan_type:
          type: string
          enum: [d_snp, i_snp, c_snp, regular]
        sponsor_name:
          type: string
        county_fips:
          type: string
        county_name:
          type: string
        state_code:
          type: string
        beneficiary_count:
          type: number
        monthly_premium:
          type: ["number", "null"]
          description: CMS monthly A/B plan premium for this county
        moop:
          type: ["number", "null"]
          description: CMS in-network maximum out-of-pocket for this county
        cost_source:
          type: ["string", "null"]
        year:
          type: number
      required:
        - id
        - plan_id
        - plan_name
        - contract_id
        - plan_type
        - sponsor_name
        - county_fips
        - county_name
        - state_code
        - beneficiary_count
        - year
    EnrollmentInput:
      type: object
      properties:
        plan_id:
          type: number
        county_fips:
          type: string
        beneficiary_count:
          type: number
        year:
          type: number
      required:
        - plan_id
        - county_fips
        - beneficiary_count
        - year
    CountyPlan:
      type: object
      properties:
        plan_id:
          type: number
        contract_id:
          type: string
        plan_name:
          type: string
        sponsor_name:
          type: string
        plan_type:
          type: string
          enum: [d_snp, i_snp, c_snp, regular]
        monthly_premium:
          type: ["number", "null"]
        drug_premium:
          type: ["number", "null"]
        moop:
          type: ["number", "null"]
        cost_source:
          type: ["string", "null"]
        star_rating:
          type: ["number", "null"]
        star_rating_part_c:
          type: ["number", "null"]
        star_rating_part_d:
          type: ["number", "null"]
        d_snp_integration:
          type: ["string", "null"]
          enum: [fide, hide, coordinated, null]
        frailty_eligible:
          type: boolean
        is_look_alike:
          type: boolean
        dual_eligible_enrollment:
          type: ["number", "null"]
        look_alike_total_enrollment:
          type: ["number", "null"]
        dual_eligible_pct:
          type: ["number", "null"]
        dual_eligible_data_status:
          type: string
          enum: [available, suppressed, unknown]
        dual_eligible_source_url:
          type: ["string", "null"]
        parent_organization:
          type: ["string", "null"]
        contract_effective_year:
          type: ["number", "null"]
        beneficiary_count:
          type: number
        year:
          type: number
      required:
        - plan_id
        - contract_id
        - plan_name
        - sponsor_name
        - plan_type
        - frailty_eligible
        - is_look_alike
        - dual_eligible_data_status
        - beneficiary_count
        - year
...